Added Autosaving
The selected settings as well as the tracker's marked items, locations and the assignment of items to locations is saved into the browser's local storage after each change.
Undo/Redo history, selected locations and currently viewed scope are lost on refreshing.

Also added a button to reset the state of the tracker.